requirements.txt
langchain
flask
flask-restful
requests
beautifulsoup4
chromadb
transformers
datasets
faiss-cpu
pyngrok
langchain-community
langchain-huggingface
python-dotenv
